requirements.txt
streamlit
openai
google-generativeai
python-dotenv
mcp 
fastapi
uvicorn
websockets
requests
PyYAML